success migrate flyway, add models, fix DDL

+++ src/main/resources/db/migration/V2__InitializeItems.sql
@@ -1,8 +1,8 @@
 CREATE TABLE if not exists items(
     item_code varchar NOT NULL PRIMARY KEY,
-    name varchar(100)
-    description varchar
-    price float8
-    stock int
+    name varchar(100),
+    description varchar,
+    price numeric,
+    stock int,
     capacity int
 );
